Выполнил задание 2. Выводим разные име, аватар, статус в зависимости от таргета
