+++ Navigation_Nebesniy/Profile line/User.swift
@@ -9,8 +9,8 @@ import UIKit
 
 final class User {
 
-    let login: String = "ACID"
-    let fullName: String = "Nebesnyi Aleksei"
-    let avatar: UIImage = UIImage(named: "Avatar")!
-    let status: String = "God please give me the strength to finish my day and hide the bodies of those that annoy me."
+    var login: String = "ACID"
+    var fullName: String = "Nebesnyi Aleksei"
+    var avatar: UIImage = UIImage(named: "Avatar")!
+    var status: String = "God please give me the strength to finish my day and hide the bodies of those that annoy me."
 }
